Refactor asset generation scripts to use shared utility module
- Extracted common functions (directory creation, material setup, mesh assignment) to asset_utils.py.
- Abstracted version-specific GeometryScript fallbacks.
- Updated all 8 biome asset generation scripts to use the new module.
- Preserved biome-specific material parameters.
