+++ scripts/asset_generation/generate_forest_assets.py
@@ -1,261 +1,78 @@
 import unreal
-import sys
-
-def ensure_directory(path):
-    if not unreal.EditorAssetLibrary.does_directory_exist(path):
-        unreal.EditorAssetLibrary.make_directory(path)
-
-def create_master_material(mat_path):
-    if unreal.EditorAssetLibrary.does_asset_exist(mat_path):
-        return unreal.EditorAssetLibrary.load_asset(mat_path)
-
-    mat_factory = unreal.MaterialFactoryNew()
-    asset_tools = unreal.AssetToolsHelpers.get_asset_tools()
-
-    pkg_path = '/'.join(mat_path.split('/')[:-1])
-    asset_name = mat_path.split('/')[-1]
-
-    mat = asset_tools.create_asset(
-        asset_name=asset_name,
-        package_path=pkg_path,
-        asset_class=unreal.Material,
-        factory=mat_factory
-    )
-
-    if not mat:
-        return None
-
-    base_color_node = unreal.MaterialEditingLibrary.create_material_expression(
-        mat, unreal.MaterialExpressionVectorParameter, -400, 0
-    )
-    base_color_node.set_editor_property("parameter_name", "BaseColor")
-    base_color_node.set_editor_property("default_value", unreal.LinearColor(1.0, 1.0, 1.0, 1.0))
-
-    roughness_node = unreal.MaterialEditingLibrary.create_material_expression(
-        mat, unreal.MaterialExpressionScalarParameter, -400, 200
-    )
-    roughness_node.set_editor_property("parameter_name", "Roughness")
-    roughness_node.set_editor_property("default_value", 0.5)
-
-    noise_node = unreal.MaterialEditingLibrary.create_material_expression(
-        mat, unreal.MaterialExpressionNoise, -400, -200
-    )
-    noise_node.set_editor_property("scale", 0.1)
-
-    multiply_node = unreal.MaterialEditingLibrary.create_material_expression(
-        mat, unreal.MaterialExpressionMultiply, -200, 0
-    )
-
-    unreal.MaterialEditingLibrary.connect_material_expressions(
-        base_color_node, "", multiply_node, "A"
-    )
-    unreal.MaterialEditingLibrary.connect_material_expressions(
-        noise_node, "", multiply_node, "B"
-    )
-
-    unreal.MaterialEditingLibrary.connect_material_property(
-        multiply_node, "", unreal.MaterialProperty.MP_BASE_COLOR
-    )
-    unreal.MaterialEditingLibrary.connect_material_property(
-        roughness_node, "", unreal.MaterialProperty.MP_ROUGHNESS
-    )
-
-    unreal.MaterialEditingLibrary.recompile_material(mat)
-    return mat
-
-def create_material_instance(mat_inst_path, parent_mat, color, roughness):
-    if unreal.EditorAssetLibrary.does_asset_exist(mat_inst_path):
-        mat_inst = unreal.EditorAssetLibrary.load_asset(mat_inst_path)
-    else:
-        mat_inst_factory = unreal.MaterialInstanceConstantFactoryNew()
-        asset_tools = unreal.AssetToolsHelpers.get_asset_tools()
-
-        pkg_path = '/'.join(mat_inst_path.split('/')[:-1])
-        asset_name = mat_inst_path.split('/')[-1]
-
-        mat_inst = asset_tools.create_asset(
-            asset_name=asset_name,
-            package_path=pkg_path,
-            asset_class=unreal.MaterialInstanceConstant,
-            factory=mat_inst_factory
-        )
-
-        if mat_inst and parent_mat:
-            unreal.MaterialEditingLibrary.set_material_instance_parent(mat_inst, parent_mat)
-
-    if mat_inst:
-        unreal.MaterialEditingLibrary.set_material_instance_vector_parameter_value(
-            mat_inst, "BaseColor", unreal.LinearColor(color[0], color[1], color[2], color[3])
-        )
-        unreal.MaterialEditingLibrary.set_material_instance_scalar_parameter_value(
-            mat_inst, "Roughness", roughness
-        )
-        unreal.MaterialEditingLibrary.update_material_instance(mat_inst)
-
-    return mat_inst
-
-def assign_material_to_mesh(static_mesh, material_instance):
-    if not static_mesh or not material_instance:
-        return
-
-    materials = static_mesh.get_editor_property('static_materials')
-
-    if not materials or len(materials) == 0:
-        new_mat = unreal.StaticMaterial()
-        new_mat.material_interface = material_instance
-        materials = [new_mat]
-    else:
-        materials[0].material_interface = material_instance
-
-    static_mesh.set_editor_property('static_materials', materials)
+import asset_utils
 
 def generate_pine_tree(mesh_path, mat_inst):
     if unreal.EditorAssetLibrary.does_asset_exist(mesh_path):
         print(f"Asset already exists: {mesh_path}")
         return
 
-    try:
-        dyn_mesh = unreal.GeometryScriptLibrary_CreateNewDynamicMesh.create_new_dynamic_mesh()
-    except AttributeError:
-        try:
-            dyn_mesh = unreal.GeometryScript_AssetUtils.create_new_dynamic_mesh()
-        except AttributeError:
-            dyn_mesh = unreal.DynamicMesh()
-
+    dyn_mesh = asset_utils.create_dynamic_mesh()
     options = unreal.GeometryScriptPrimitiveOptions()
+    primitive_functions = asset_utils.get_geometry_script_primitive_functions()
     transform = unreal.Transform()
 
     try:
-        if hasattr(unreal, "GeometryScriptLibrary_MeshPrimitiveFunctions"):
-            unreal.GeometryScriptLibrary_MeshPrimitiveFunctions.append_cylinder(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=transform,
-                radius=15.0,
-                height=150.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
+        primitive_functions.append_cylinder(
+            target_mesh=dyn_mesh,
+            primitive_options=options,
+            transform=transform,
+            radius=15.0,
+            height=150.0,
+            radial_steps=6,
+            height_steps=1,
+            capped=True,
+            origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
+        )
 
-            leaf_transform = unreal.Transform()
-            leaf_transform.translation = [0, 0, 80.0]
-            unreal.GeometryScriptLibrary_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=60.0,
-                top_radius=0.0,
-                height=150.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
+        leaf_transform = unreal.Transform()
+        leaf_transform.translation = [0, 0, 80.0]
+        primitive_functions.append_cone(
+            target_mesh=dyn_mesh,
+            primitive_options=options,
+            transform=leaf_transform,
+            base_radius=60.0,
+            top_radius=0.0,
+            height=150.0,
+            radial_steps=6,
+            height_steps=1,
+            capped=True,
+            origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
+        )
 
-            leaf_transform.translation = [0, 0, 140.0]
-            unreal.GeometryScriptLibrary_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=50.0,
-                top_radius=0.0,
-                height=120.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
+        leaf_transform.translation = [0, 0, 140.0]
+        primitive_functions.append_cone(
+            target_mesh=dyn_mesh,
+            primitive_options=options,
+            transform=leaf_transform,
+            base_radius=50.0,
+            top_radius=0.0,
+            height=120.0,
+            radial_steps=6,
+            height_steps=1,
+            capped=True,
+            origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
+        )
 
-            leaf_transform.translation = [0, 0, 190.0]
-            unreal.GeometryScriptLibrary_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=35.0,
-                top_radius=0.0,
-                height=100.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
-        else:
-            unreal.GeometryScript_MeshPrimitiveFunctions.append_cylinder(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=transform,
-                radius=15.0,
-                height=150.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
-
-            leaf_transform = unreal.Transform()
-            leaf_transform.translation = [0, 0, 80.0]
-            unreal.GeometryScript_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=60.0,
-                top_radius=0.0,
-                height=150.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
-
-            leaf_transform.translation = [0, 0, 140.0]
-            unreal.GeometryScript_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=50.0,
-                top_radius=0.0,
-                height=120.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
-
-            leaf_transform.translation = [0, 0, 190.0]
-            unreal.GeometryScript_MeshPrimitiveFunctions.append_cone(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=leaf_transform,
-                base_radius=35.0,
-                top_radius=0.0,
-                height=100.0,
-                radial_steps=6,
-                height_steps=1,
-                capped=True,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
+        leaf_transform.translation = [0, 0, 190.0]
+        primitive_functions.append_cone(
+            target_mesh=dyn_mesh,
+            primitive_options=options,
+            transform=leaf_transform,
+            base_radius=35.0,
+            top_radius=0.0,
+            height=100.0,
+            radial_steps=6,
+            height_steps=1,
+            capped=True,
+            origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
+        )
     except Exception as e:
         print(f"GeometryScript generation failed: {e}")
         return
 
-    create_options = unreal.GeometryScriptCreateNewStaticMeshAssetOptions()
     try:
-        if hasattr(unreal, "GeometryScriptLibrary_CreateNewStaticMeshAssetFromDynamicMesh"):
-            static_mesh = unreal.GeometryScriptLibrary_CreateNewStaticMeshAssetFromDynamicMesh.create_new_static_mesh_asset_from_dynamic_mesh(
-                dynamic_mesh=dyn_mesh,
-                asset_path_and_name=mesh_path,
-                options=create_options
-            )
-        else:
-            static_mesh = unreal.GeometryScript_AssetUtils.create_new_static_mesh_asset_from_dynamic_mesh(
-                dynamic_mesh=dyn_mesh,
-                asset_path_and_name=mesh_path,
-                options=create_options
-            )
-
-        assign_material_to_mesh(static_mesh, mat_inst)
+        static_mesh = asset_utils.create_static_mesh_from_dynamic_mesh(dyn_mesh, mesh_path)
+        asset_utils.assign_material_to_mesh(static_mesh, mat_inst)
     except Exception as e:
         print(f"GeometryScript mesh bake failed: {e}")
 
@@ -265,64 +82,31 @@ def generate_forest_rock(mesh_path, mat_inst):
         print(f"Asset already exists: {mesh_path}")
         return
 
-    try:
-        dyn_mesh = unreal.GeometryScriptLibrary_CreateNewDynamicMesh.create_new_dynamic_mesh()
-    except AttributeError:
-        try:
-            dyn_mesh = unreal.GeometryScript_AssetUtils.create_new_dynamic_mesh()
-        except AttributeError:
-            dyn_mesh = unreal.DynamicMesh()
-
+    dyn_mesh = asset_utils.create_dynamic_mesh()
     options = unreal.GeometryScriptPrimitiveOptions()
+    primitive_functions = asset_utils.get_geometry_script_primitive_functions()
     transform = unreal.Transform()
 
     try:
-        if hasattr(unreal, "GeometryScriptLibrary_MeshPrimitiveFunctions"):
-            unreal.GeometryScriptLibrary_MeshPrimitiveFunctions.append_box(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=transform,
-                dimension_x=120.0,
-                dimension_y=100.0,
-                dimension_z=50.0,
-                steps_x=3,
-                steps_y=3,
-                steps_z=3,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
-        else:
-            unreal.GeometryScript_MeshPrimitiveFunctions.append_box(
-                target_mesh=dyn_mesh,
-                primitive_options=options,
-                transform=transform,
-                dimension_x=120.0,
-                dimension_y=100.0,
-                dimension_z=50.0,
-                steps_x=3,
-                steps_y=3,
-                steps_z=3,
-                origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
-            )
+        primitive_functions.append_box(
+            target_mesh=dyn_mesh,
+            primitive_options=options,
+            transform=transform,
+            dimension_x=120.0,
+            dimension_y=100.0,
+            dimension_z=50.0,
+            steps_x=3,
+            steps_y=3,
+            steps_z=3,
+            origin=unreal.GeometryScriptPrimitiveOriginMode.BASE
+        )
     except Exception as e:
         print(f"GeometryScript generation failed: {e}")
         return
 
-    create_options = unreal.GeometryScriptCreateNewStaticMeshAssetOptions()
     try:
-        if hasattr(unreal, "GeometryScriptLibrary_CreateNewStaticMeshAssetFromDynamicMesh"):
-            static_mesh = unreal.GeometryScriptLibrary_CreateNewStaticMeshAssetFromDynamicMesh.create_new_static_mesh_asset_from_dynamic_mesh(
-                dynamic_mesh=dyn_mesh,
-                asset_path_and_name=mesh_path,
-                options=create_options
-            )
-        else:
-            static_mesh = unreal.GeometryScript_AssetUtils.create_new_static_mesh_asset_from_dynamic_mesh(
-                dynamic_mesh=dyn_mesh,
-                asset_path_and_name=mesh_path,
-                options=create_options
-            )
-
-        assign_material_to_mesh(static_mesh, mat_inst)
+        static_mesh = asset_utils.create_static_mesh_from_dynamic_mesh(dyn_mesh, mesh_path)
+        asset_utils.assign_material_to_mesh(static_mesh, mat_inst)
     except Exception as e:
         print(f"GeometryScript bake failed: {e}")
 
@@ -331,11 +115,11 @@ def main():
     base_dir = '/Game/Art/Models/Forest'
     mat_dir = '/Game/Art/Materials'
 
-    ensure_directory(base_dir)
-    ensure_directory(mat_dir)
+    asset_utils.ensure_directory(base_dir)
+    asset_utils.ensure_directory(mat_dir)
 
     master_mat_path = f"{mat_dir}/M_StylizedForest"
-    master_mat = create_master_material(master_mat_path)
+    master_mat = asset_utils.create_master_material(master_mat_path)
 
     pine_mat_path = f"{mat_dir}/MI_PineTree"
     pine_mat = create_material_instance(pine_mat_path, master_mat, [0.05, 0.25, 0.05, 1.0], 0.8)
